Add files via upload
The Phase II clinical trial dataset for a new oral insulin called Auralin is used for this project.
DISCLAIMER: This Data Isn't "Real"
The Auralin and Novodra are not real insulin products. This clinical trial data was fabricated for the sake of this project. When assessing this data, the issues detected (and later cleaned) are meant to simulate real-world data quality and tidiness issues.

That said:
*This dataset was constructed with the consult of real doctors to ensure plausibility. *This clinical trial data for an alternative insulin was inspired and closely mimics this real clinical trial for an inhaled insulin called Afrezza.

The data quality issues in this dataset mimic real, common data quality issues in healthcare data. These issues impact quality of care, patient registration, and revenue.
The patients in this dataset were created using this fake name generator and do not include real names, addresses, phone numbers, emails, etc.
Library need to be installed :
pandas